Streamline the Zot Clock
(Alternate commit title: Zot Around The Clock)

The Zot clock has been working OK as a time limit, but
it has a major shortcoming: it's not explained to players
at all. The reasoning there was that we didn't want to
stress players out or to make it easier to sneak around
the clock, but the very loose time limits on the current
clock make it easy to sneak around regardless for a spoiled
player, and unexplained mystery deaths are more stressful
than occasional messages.

(Also, if players don't know that the clock exists, it
doesn't serve its intended role of discouraging very slow
play!)

So, derandomize the Zot Clock, remove the drain penalty,
display remaining turns when there are <1k left, and add
an option (off by default) to display remaining turns at
all time. Also, take another pass on the messaging. That
should hopefully leave the whole thing in good enough
shape for release.