A themed vault with boulder beetles
In gammafunk_indiana_jones, you find the corpse of adventurer in a
narrow, spider-filled corridor leading to a "trap" room with a boulder
beetle (or two or three), that further leads to a trapped treasure room.
In earlier depths, this is flame clouds in front of lava snakes
(snakes...why did it have to be snakes...), and later on this becomes a
room with fast moving spiders and Zot traps. If the player fights the
beetle(s) in the long corridor, they're fairly likely to get flattened
by a rolling bug.

The loot consists of the characteristic whip and hat near the corpse,
which scales in quality and with the whip eventually becoming a demon
whip, along with jewellery plus possible gold in the treasure room. The
monsters and traps likewise scale in number and dangerousness with
depth, increasing the number and dangerousness of spiders as well as the
number of boulder beetles.

It'd be nice to better decorate this vault in Tiles with a more forest
temple theme, but there aren't very appropriate stone wall tiles for
this purpose.