A boulder beetle transporter vault
In gammafunk_boulding_lane, the player transports into the far side of
the vault with one to three boulder beetles at the far side. If you can
manage to get in before the boulder beetles all see you, you might avoid
their rolling down the lanes as they look to mark a strike. The loot
always contains a pre-identified pair of boots of rampaging (which can
be barding). The number of beetles and loot quality and quantity scales
with depth, with up to 3 beetles appearing along with an eventual torpor
snail. The loot scales up to artefact quality boots and another
star_item.