Remove references to globals across files
I found three separate instances of functions reusing a global variable
exposed by travel.cc for their own scratch space, all for mild
variations on the same BFS algorithm. This commit replaces that with a
local variable, and one of a more sensible type as well; only a boolean
visited / not visited state needs to be tracked.

I don't quite have the energy to properly extract these right now, but
these would possibly be a good candidate for unit testing.