Add transporter vaults to Vaults
With c7717499, we can now place transporter vaults in Vaults. This
commit adds such vaults to this branch as "hard" vault rooms. The
structure of Vaults branch room vaults is such that vaults often need
substantial internal changes in order to be suitable. That usually
requires a separate vault definition, but these transporter vaults work
directly as vault rooms, so I've just given them the necessary tags.

For some vaults, I've added a border of floor around the vault that is
removed when not placing in Vaults branch. I experimented with simply
not having this, which would have allowed the vaults to connect two
otherwise disconnected areas, as they can do when placing outside
Vaults, but the vaults sometimes placed in rooms so that one side was
disconnected from the rest of the map. The necessary "passable" tag
disables checks against this that would otherwise normally generate a
veto, so I'm forced to add this path border. Ideally the zone validation
would know how to path through transporters so we wouldn't need
"passable" at all, but mumra's approach works for now.

For vaults with thematic monsters, I made appropriate choices that are
usually a mix between late Dungeon and Depths, since Vaults shares a lot
in common with both areas. Generally all monsters used can already place
in Vaults, since this is a stipulation these vaults already use for each
branch.