Adjust doors in a runed door vault
Make the entire entrance to gammafunk_indiana_jones runed to avoid
autoexploring into a Zot trap, and to generally improve the structure of
the vault. Keep the "trap" gate as a (non-runed) clear door, but the
final door normal, since the glass around it takes care of visibility.
Also make the glass stone instead of rock.