refactor: Remove declarations of internals in .h
This commit removes the declarations of _setup_fixture_lua
and _teardown_fixture_lua functions from the fixture_lua.h
header file. This is because the functions are
implementation details. In test_fixture_lua.cc, the commit
replaces direct calls of these functions by automatic
calls to the destructor of the fixture.