test: Use random coordinates to test FixtureLua
Since setup and tear-down of the fixture require
considerable time, the tests only use a very limited number
of coordinates. This commit adds a class to randomize the
coordinates to enhance the chance to discover potential
problems without using more time. The class generates
x- and y- coordinates that are uncorrelated.