Split debug defines out of AppHdr.h
These defines sync DEBUG and NDEBUG. Normally, I'd do this in the
Makefile, which would ensure that all compilations have only one of
these defined, but unfortunately different compilers provide different
options.

Instead, we can move this to its own header file. This allows it to be
included from debug.h, which depends on these rules to verify that they
are not out of sync. Really, all .h and .cc files should include this,
as header files may define static functions which contain asserts, which
would otherwise invoke the assert macro before possibly defining NDEBUG.

Probably the ideal solution in the general case is to have a single
header file that all other headers must include. This should be
extremely minimal, containing only definitions that are vital for
correctness, such as this DEBUG/NDEBUG syncing. Specifically,
convenience definitions should not be included: AppHdr.h is already a
mess in this regard.