Fix ccache not working on GitHub Actions CI
ccache was not working for a number of reasons. First, the compiler was
not being invoked through ccache: this requires invoking the compiler
via a symlink named after the compiler binary that points to the
ccache binary.. An alternative method is to add a directory containing
preprepared symlinks to the PATH. alexjurkiewicz/setup-ccache does this,
but it seems the path it prepends is incorrect.

Next, GHA will not actually *update* the cache contents if the cache
was restored with a direct key match! In order to fool GHA into doing
so, we append the SHA1 of the commit being tested, ensuring that each
cache save is saved with a unique key (jobs with the same SHA are
disambiguated by matrix parameters). Cache restoring therefore relies
on partial matches, and will restore the most recently saved cache
snapshot.

However, GHA cache restore prefix matching apparently requires a dash
at the end of each pattern in order to match. This isn't documented
explicitly (of course), but all their examples use this format, and
adjusting this caused GHA to correctly restore the most recent cache.

This commit also zeroes the ccache stats before running, so that the
stats printed at the end reflect only the current compilation job. If
ccache breaks in the future, this will ensure that we aren't fooled by
the cache being restored, but never used, and showing old stats.

With high hit rates, this roughly halves runtime, from ~15 minutes to ~6
minutes. Major remaining sources of latency:
 - ~3 min: collecting coverage with lcov.
 - Install requirements/dependencies: ~1.5min for gcc, 2min for clang.
 - actions/checkout: ~40sec.
Of these, installing requirements/dependencies can probably be optimized
by caching the package manager cache directories.