Add an option which means autoexplore/travel treat closed doors more like walls than floor.
Previously, travel_open_doors had two values.
If true, the travel command would open a door if it was quicker than going around. The explore command would open the door if this was the quickest way to get to something interesting.

This adds a third in which it treats it in the same way it treats a runed door.

Travel takes a route which avoids doors if it can do so. If not, it walks up to a door and stops.

I've based the option setting code on easy_confirm. It's a little more complex, as travel_open_doors is exposed to Lua. I used a macro for the list of values  the option can take to avoid having to have more than one copy of the list in the source code.