Improve handling of builder failures with pregen
Before pregen, a builder failure would lead to a force save, and a new
attempt to generate the failed level on load. After pregen, something
similar would have most likely happened, but on load the player would (I
think) be teleported to the failed level and pregen would break. At some
point I added an ASSERT to the save code for other reasons that would
trigger on this case, so most recently the builder would crash without
saving, leading to a broken save (but that could be played up to a
point).

This commit restores a version of the original behavior. If the builder
fails, the game will force-save and then crash. The save step discards
the builder's attempts, but keeps the effect on the rng, so that the
level seed will be different on the next attempt, and most likely
succeed. The end result is still deterministic overall given a starting
seed. When the player loads the save, the builder picks up where it died.
I have been testing this in trunk with seed 17783394655777042589, which
triggers a builder failure (50 vetoes in a row) on depths 3.

This commit also contains a small improvement to the messages API which
didn't end up being needed for the final implementation, but is still a
useful addition.