Attempt to fix contrib zlib on mac gha builds
This adds a missing header file that is now triggering errors instead of
warnings on this build. The crawl-zlib repository is in a somewhat
confusing state, so while it might be nice to update the zlib version, I
was not able to figure out how to do that.

(cherry picked from commit 95a0c6c400634562f1310753810b201acce95f69)