Lower a polynomial for linewrapping in local tiles
When looking for lines that exceeded the wrap width, this code, for each
line, unnecessarily scanned through the entire remaining string, doing
glyph width calculations the whole way. Needless to say this made
linewrapping (and therefore Text rendering) quite a bit slower for very
large text boxes in tiles. This change only checks for this case on the
current line.

(Even with this change, Text rendering speed is not acceptably fast for
large textboxes in tiles, e.g. from builder output in a debug build, so
something else is still going on.)