Add a rudimentary veto debugging mode
This mode modifies the builder so that, after a wizmode travel command
or &ctrl-r (reload level), instead of retrying on a veto, it exits
immediately. The effect is that you can look at the partially completed
level, and together with the builder log (&ctrl-l on a fulldebug build),
get a sense for why the veto happened. This is implemented with stable
seeding, so that level reload should resume exactly where the builder
left off.

This is implemented via an #ifdef because the game isn't good for much
else in this mode and anyone who is using this is running a fulldebug
build and so can rebuild anyways, but it could in principle be
implemented with a more user-friendly interface as an option or
something like that.