Increase broad axe delay
It's a little too easy to get this rather strong weapon online.
Comparing to, for example, a battleaxe, the broad axe gave the ability
to use a shield for only -2 base damage and *less* skill required to
reach mindelay!

That sort of dynamic is fine when we're comparing rare weapons to
common ones (e.g. double sword vs great sword, eveningstar vs great
mace), but broad axes are quite common. Per objstat, there are about
13 broad axes in the average 3 rune game, with perhaps 1 appearing in
D, 1 in Lair, etc. That's over 13 times as common as the rare 1h weapon
types, and appearing much earlier.

It's possible that there should be a comparable split for axes, with a
'common' and 'rare' 1h axe in the same way that we have scimitars and
double swords or morningstars and eveningstars. For now, though,
increasing the skill requirement for broad axes seems like the best way
to keep them a viable endgame weapon while making other weapons a bit
more competitive.