Revamp a Shoals ghost vault
The vault ebering_ghost_davey_jones places a kraken (or two derived
undead krakens) and a ghost, but it's trivial to separate the
ghost-kraken alliance. Even with these two monsters together, the vault
would be pretty easy. This commit revamps the vault to have a
transporter, adds some extra critters, and reworks the themeing.  Now
only one kraken is ever placed, either normal or spectral, and there are
accompanying monsters that match the normal/spectral themes.

We now have deep sea creatures that an unfortunate seagoer who ended up
in Davey Jones' locker might find, such as sea snakes, snapping turtles,
and, rarely alligator snapping turtles (we'll pretend the turtles are
sea turtles). There's also a merfolk with a nice weapon and a monster
band. For the natural theme, this is an aquamancer and water elementals.
For the spectral theme, this is a merfolk impaler where the impaler and
its band all have draining weapons. The critters for the spectral theme
are spectral versions of their natural counterparts.

The loot has been somewhat buffed, and the internal layout is now
symmetrical, but with some added walls that have a couple pattern
variants. This new version of the vault should be nasty enough to get a
few kills while hopefully offering some enticing rewards.