Add a script for automated placement and connectivity testing
This repurposes an ancient, and dormant, test that placed a single vault
and adds a bunch of features to make it useful for bulk testing vaults,
targeting basic placement issues and connectivity. This is extremely
useful for finding rare connectivity issues, and for doing a pass
through all vaults to find obvious veto-triggering closets.
Unfortunately I suspect it takes to long to be viable as a CI test
(though I may still try this out), and is better interactively. Initial
fixes as a result of developing this script will come in separate
commits. Use case:

```
$ util/fake_pty ./crawl -script placement.lua minmay_nonomino_d4 -count 10 -dump
Testing map 'minmay_nonomino_d4'
    Failing vault output to: placement-minmay_nonomino_d4.10.txt
script error: ./scripts/placement.lua:184: Isolated area in vault minmay_nonomino_d4 (2 zones)
```

This generates this vault on a blank level 10 times and outputs each map
to a file. In this case, it failed on try 10 (which aborts) and so it
stops then and points out the specific failing file to you. This file
shows the whole level, but here's an excerpt illustrating the vault
itself:

```
```

In this case the problem is pretty visible, though sometimes
connectivity issues do not pop out in the test output (e.g. ones
involving water, or CLEAR tiles, etc.) and need further debugging.

Currently this script is geared towards finding isolated areas that will
trigger vetoes, though it might be extendable to do other related
things that it doesn't know (e.g. find tele closets in opaque maps).

The script skips various kinds of maps, and is hard-coded to skip
certain specific maps that either are ones I haven't yet figured out how
to solve, or illustrate some other unusual circumstance that can't be
detected from tags. The two biggest skipped categories are encompass
maps, and unrand maps. The latter are usually pure subvaults, so testing
the map that places them exercises them. (Though in practice, you need
to manually test these vaults with large -count arguments. I've done
this in a few cases, e.g. minmay_the_grid maps, but not generally.)