Tweaks to minmay_the_grid for occasional closets
These closets are of the veto-triggering kind, not the tele closet kind.
I'm not very confident that these are the ultimately best fixes, but
they are fixes. They target an issue with a specific style of subvault
that relied on the surrounding tiles for connectivity: these subvaults
assume external passable squares that lined up with the internal gaps,
but it's possible for those external squares to become walls, leading to
closets.

Otherwise I've done a lot of bulk testing of minmay_the_grid vaults, and
haven't found other connectivity issues. (Right now my script can't
place the ultimate one, though, so that could use some followup.)