Clean up plants in post-placement fills (Lightli)
Regular monsters (and items) shouldn't appear outside of vaults at this
point in the level generation process, but plants are entirely possible
in some layouts. This commit allows plants to be cleaned up, and
prevents post-placement filling of squares with non-firewood. (The fill
code itself will wipe out anything, beware.)