Allow quivering evokable unrands
There are only three of these, and two are not common at all, but it's
actually pretty useful for olgreb. Because of the unrand API, this
implementation is a bit hacky, but with only a few items I'm not
currently excited about writing a more general API for some of this
stuff.