Turn monster explosions into fineffs
Explosions firing immediately causes a pretty classic set of issues -
we're invalidating the old state of the world in surprising ways. For
example, if a ball lightning explodes and kills an attacker midway
through the resolution of a multi-target attack (e.g. a javelin toss),
the game just has no way to figure out how to resolve the rest of the
attack afterward - it can't generate new ranged_attacks without the
original attacker, so it just crashes.

Instead, let's wait until the *end* of any ongoing melee attacks or
beams before triggering death explosions. The timing here is slightly
different for attack-triggered fineffs - we trigger them a little later
now melee attacks, in fight.cc - but hopefully that's ok.

Overall, this code is very bad and probably buggy. I'm sorry.