Simplify and document the actions/quiver API somewhat
Mixing shared_ptr and references was too confusing, with some awkward
pitfalls. Given that the whole point of action_cycler is to ensure
non-null, this commit just switches things over to use only shared_ptrs,
and makes explicit this (and cases where nullptr may be returned) in the
doc comments.

Also, a few other tweaks/fixes that I noticed along the way.