Handle randomization correctly in absolute zero targeter
Before, I had assumed that this wasn't randomized. However, it is,
inside distance_iterator. So, this commit changes the targeter to show
AFF_MAYBE on the closest targets when there is more than one at the same
distance. The visual difference is a bit subtle, so it may be that this
will be confusing, but this still seems the best way to handle it.