Rework player spell accuracy display
It didn't work - the approximation was too crude. We pretended we
were only rolling a random number once (for to-hit), but we actually
rolled target EV 3 times as well - and more rolls if the target had
rMsl!

We now try to simulate most of that (though not rMsl). It seems like
even in the worst likely cases (high-accuracy spells against a haloed
high-EV target), we only run about 150k iterations, which is very fast
locally. This produces much more accurate results.

The correct solution is to simplify the beam accuracy calculations,
as we have with melee/ranged attack accuracy, but that's a larger
task than I want to attempt this close to feature freeze.