More spell quivering fixes, this time for excursions
81e999faa9 was specifically triggered by checking spell uselessness of a
quivered spell while doing levelgen. Crashes can also happen if this is
checked during an excursion. Before, none of this code would have a way
of being called, but when a spell is quivered on newquivers, this code
is used to figure out whether to gray out the spell, so it is triggered
by ui redraws as well as sending player info in webtiles. I've added
another sanity check to `spell_no_hostile_in_range`, but also just
headed this off at the outset in `spell_uselessness_reason`.