Generate a quickstart.txt
23cac7494c converted the quickstart guide to markdown, but at the time
of merge I don't think anyone fully realized that this was loaded
in-game. This fact has caused a bunch of havoc.

This commit generates a .txt by simply copying the md file, which should
eventually be substituted for better rendering. Two reasons for doing
this: (i) for in-game use, a .txt version would be better since we don't
have any markdown rendering in-game, and (ii) on some servers (perhaps
just CWZ) the .md file is not getting copied into the game data
directory, which is causing a ton of crashes. This could be fixed on the
server, but given (i) it also seems reasonable to fix it on our end.