Prevent most quiver autofilling
This commit effectively restores oldquiver behavior: running out of ammo
will autoquiver other ammo using the settings in fire_order, but nothing
else. The only other action type that can run out is wands, and these
result in an empty quiver. Dropping items will also lead to an empty
quiver, which I'm less sure of.

This commit is likely not the last word on quiver autofilling, but it
does solve one of the more annoying sets of interactions, which is when
running out of ammo and hitting autofire triggers something extremely
different, such as invoking berserk.

Because wands can no longer be accidentally triggered by running out of
ammo (see f7e0b5f77) I've added them back in to the cycle order.