Shift nrook's guardian serpents deeper
D:9 is a bit earlier for guardian serpents. So is d:11, really, but
what's life without a *little* excitement?

nrook_guardian_serpent_double might be too spoilery to exist (unrelatedly)
but I don't have the heart to change it. (And as spoilery vaults go,
the trick in this one is pretty obvious.)