Add an admin panel tool for password resets
For servers that can't set up password resets by email, this allows
admins to manually generate reset tokens. I implemented it because
handling manual password resets on CAO is extremely annoying, but
setting up the email mode is unlikely to happen.

This feature is very much off by default because it's a drastic increase
in privileges for the way admin users are set up on many servers --
right now, the admin flag is mainly for identifying who can do rebuilds
/ use wizmode, i.e. almost nothing that can affect other users. So I
don't think server owners will obviously want to enable a way of
changing arbitrary passwords (if you have the reset link, you can change
the password) for anyone who can rebuild. Really this calls for multiple
tiers of permissions, but that's pretty involved to figure out how to
do. TBD.