CLI for setting password reset tokens
I had qualms about enabling the admin panel version of this on a server
with a very long admin list, so in lieu of adding a second DGL
permissions tier (for now) I'm adding a CLI version of this to simplify
password resets a little bit. Not yet tested in a chroot setting.

This change enables servers accepting reset tokens unconditionally; but
unless one of the other two options is enabled they can still only be
set via the command line.