Add a quiver swap command
This new command swaps between the current quiver action and the most
recent valid quiver action. Default bound to `]`.

Since quiver actions don't have stable lettering, it's a bit hard to see
exactly how to do this, but I wonder if something more controllable
along the lines of `'` for weapons might eventually be in order?