Rename magic resistance to willpower
Magic resistance is a stat that covers a very wide
gamut of effects, from hexes (Slow, Confuse) to
vaguely hex-like things (Teleport Other, Siren Song)
to things that very clearly aren't hexes at all
(Disintegrate). Of course, it doesn't cover other
hexes, like Metabolic Englaciation and Dazzling Flash,
and it definitely doesn't cover all 'magic' - it
provides no resistance to a magical fireball or a
magic sword.

In short, it's a name that's both unspecific and
*too* specific. It actively misleads players. It has
negative value!

It's impractical and undesireable to change the game's
systems to match the literal meaning of 'magic resistance',
and there's no more precise term that matches the actual
function. You could say something like 'hex resistance',
but again, you run into the issue of hexes that aren't
resisted by hex resistance and non-hexes that are resisted
by hex resistance.

So, instead, let's rename magic resistance to 'willpower'.
This is a vaguer term, which means that it shouldn't
mislead players in the same way, but it's entertainingly
flavourful - it's fun to talk about Trog's influence
making you strong-willed or Ru's sacrifices making you
weak-willed. (Credit to Midn8 for "force of will"
inspiring this.)

One weird note about this name is that Disintegrate will
be willpower-resisted. I've added a cute little bit of
flavour to the item description as justification, and
that wand might not be long for this world in any case,
so it seems fine.

There's no save compat code here, so this is easy to
revert or tweak further as needed.