Initialize sockaddr_un before initial recvfrom
This is a weird one. Under some circumstances that I cannot fully
determine, the socket address in srcaddr after the recvfrom call is not
getting 0-terminated, leading to an invalid pathname (with garbage
characters after the end of the actual file name) and ENOENT on the
first sendto immediately after the attach message. This leaves the crawl
process with no listening clients; the way this appears from the user's
end is that webtiles just hangs on the loading screen. Memsetting the
structure before recvfrom fixes it, and is good practice anyways.

*However*, this seems to be Tornado version specific: I get this issue
almost deterministically on Tornado 6.1.0, often on Tornado 6.0.4, and
sporadically on 6.0.3 (but I have seen it happen). This makes me suspect
that there is some kind of race condition where the socket may not be
set up on the client end quite in time for the recvfrom call, and this
weird behavior is just a symptom of the race condition. However, it
isn't at all obvious why this could happen on the python side, and
there's no async code I can detect that would be interfering. So I'm not
really sure what is going on here.

It's also possible this issue, or at least the particular symptom I'm
fixing here, is mac specific, since I think we would have heard about
this if one of the tornado 6 servers were having this problem.