add a new overflow vault for gozag and zin
This vault is an idea inspired by zin_treasury that I thought
could be an amusing play on the contrast between Zin and Gozag.
Gozag tempts the player with a single-charged wand of digging
to steal a negligible amount of gold from Zin's treasury just for
the sake of it.