Darken out of range tile
The range distinction was extremely subtle on many dungeon backgrounds,
because tiles shows both range and los in the targeter. This makes the
range distinction much sharper, at the cost of making the oor vs oos
distinction fairly subtle. However, this isn't very important in the
targeter, so I think this is probably an improvement.

True confessions: I couldn't find a good free sprite editor that could
correctly handle the alpha channel in this file, so this is actually
just the old out_of_sight.png from before 010c0ec4c0e. I kept the
lighter version around in thee repo in case someone wants to play around
with this in more detail. However, we really should be doing this kind
of darkening programmatically.