Improve autofight_throw and CMD_AUTOFIGHT_NOMOVE
The main content of this change is to allow autofight commands (as
opposed to CMD_AUTOFIRE) to only trigger actions that do direct damage
of some kind. Several people have pointed out to me that their behavior
with something like Berserk quivered, vs. how players used them before
quivers, was not very coherent. Now, any autofight command will act as
if the quiver is empty when an action that doesn't do direct damage is
quivered. This will also hopefully make the quiver more usable when
CMD_AUTOFIGHT_NOMOVE is used as the Tab binding.

As a bonus, autofight_throw=true now correctly handles the case where a
spell is out of range, moving towards the nearest target or falling back
to melee depending on range. This is fairly useful for many spells, but
especially for something like hailstorm.

There was no existing code to make this distinction, so I mostly ended
up hardcoding it. Unfortunately this may mean that there are missed
cases.