Add some sanity checks to IMB
The effect of this spell when one of the actors was knocked into a
dispersal trap was fairly odd, and these changes make it slightly less
odd. (They may or may not interact with crashes we've been seeing.) It's
possible that for knockback specifically, dispersal should act as a
fineff, but this isn't a good approach for dispersal in general.