Fix a thunderbolt crash
Thunderbolt was not correctly handling the case where a player moved
instantaneously or during the firing turn into the position that they
had just targeted, resulting in a ray crash on a length 0 ray. This
could happen easily in wizmode but also in a few ways during a game,
including mesmerise, serpent's lash, and probably knockback. This commit
guards against the specific case, though not for non-crashing cases of
the player moving in one of these ways (which might lead to bizarre
thunderbolt shapes). For fully correct behavior, it should probably
check not just the turncount of the previous firing, but the position
the actor was at at the previous time of firing.