Don't do a weapon check on autofight movement for launchers
When an enemy was visible but fire was blocked, a wielded launcher would
trigger a "really attack while wielding" message before allowing
movement towards the enemy. This old behavior was strictly in keeping
with the documentation for autofight_prompt_range, but I think still
unexpected, and with confusing messaging (referencing attacking) on top
of that. So just skip the wielded weapon check altogether if wielding a
launcher. This does obscure corner cases where a launcher has an
inscription, but since all launchers have full range, it's pretty
restricted.

Resolves #1614