Actually stop the Zot clock when it should be paused or stopped
The Zot clock ticks even when in the Abyss, during the orb run, and in
Sprint mode. It cannot kill the player, but `turns_until_zot()`
showing a negative number of turns left in `E` looks like a bug.

Also, this commits tweaks the messaging about the Zot clock in `E`.