Try to fix the renounce to `X` mapping (10637)
This can happen because `your_talents` is extremely side-effect-y, and
if it can get called before `set_god_ability_slots`, it will actually
set lettering for whatever god abilities appear to be available at 0
piety. There are various ways that this might happen, centering around
the messaging on join. I suspect the original version of 10637 involved
local tiles (where the ability buttons are constructed by calling
`your_talents`), but the modern version of this is triggered by having
an ability quivered when joining the god; a redraw on the messaging also
calls `your_talents` when checking quiver enabled status. The best fix
would be a more sane refactoring of how to check if the player has an
ability, but this quick fix should prevent this particular problem.

I won't be too shocked if this leads to crashes, because of weird join
religion ordering effects; this is not unrelated to the crash fixed by
ca8e9911413a.