Remove the anti-cannibalism conduct (12368)
With eating gone it looks weird and out of place. It still had an effect
in what one could devour with hydra form while worshipping a good god. A
lot of code for a small effect.

Beoghites still don't eat fellow orcs, on accounnt of the corpse
violation conduct and also the practical desire of preserving an orc
corpse to resurrect them.