Fix ghouls not being able to use potions (PR 1642)
The removal of the notion of "hungry undead" didn't update the logic
that assumed all "pure" undead can't use potions. Such undead were
previously only mummies or temporary lich form, but now it includes
ghouls. This commit fixes the logic in the obvious way for now, but the
undead-state-type likely enum needs a refactor, as its not doing a great
job of giving proper undead knowledge at present.

I'm not working with the referenced PR given the source, but will leave
it open should anyone else wish to.