More WL -> Will (Goratrix)
Consistent terminology for player-facing messaging.

I'm intentionally leaving the player -> monster hex targeter with a
different message ('chance to affect') than the monster -> hex messaging
('chance to defeat your Will'). The latter feels quite awkward in the
targeter, and I think it's maybe OK to be a little less specific on what's
going on under the hood. We'll see!