Prompt before ending Ozocubu's Armour (gammafunk)
From conversations with gammafunk and also other players and my own
experience, my cancellation of Ozo's armour by movement is more often
inadvertent than an intentional decision. Add some ui ergonomics for
this.

I didn't add an rc option for whether or not to prompt: we already have
the answer prompt lua callback, and from that players who want the
prompt auto-answered one way or the other (or in a context-sensitive
way) can configure the behavior to their hearts delight.