Make frozen ramparts end on movement
Frozen Ramparts' combination of damage over time and area of effect made
it one of the most powerful single school L3 spells. On its own, this
isn't necessarily a bad thing as it is not overpowered as a L3 spell.
However, FR encourages motion, which is an anti-synergy with another low
level Ice spell: Ozocubu's Armour.

This tension combined with FR's high power-level and the average overall
high power-level of Ice spells motivates changing FR to end on movement
(the exact rules are the same as Ozocubu's armour: blinking and getting
trampled don't end it).