Give the cloak of Starlight a dazzling effect (amcnicky)
This is a different take on the submission in #1572 to make the cloak
of Starlight more interesting than a cloak with a few artefact properties and
a nice tile.

This commit removes the resist stack, but keeps the stealth malus and
EV.

In exchange it now grants a passive "Dazzle" effect: on one out of 5
dodges it attempts to blind the attacker, using the same code path and
roll as dazzling flash at power 100. This puts it in contrast to the
other unrand cloaks which grant their passive effect on hit.

Closes #1572