Improve a corner case for `F`
When `F` is pressed with no throwables, the game opens the full
inventory. This seems (maybe) sensible, but it doesn't really cue that
it's done anything out of the ordinary, so this commit removes the
special case from invent.cc and changes the menu prompt to try to
indicate this.