Revise the sword of Power (kate-)
The variable plus of the old SoP encouraged swapping at low health, and
while getting a giant plus (+27!) could be fun with berserk or Ely, the
lack of a brand and declining plus at lower health made it not that
attractive.

Instead, the sword becomes a fixed plus +5 vorpal greatsword that fires
a "sword beam" reminiscient of the old bolt of inaccuracy on a chance
depending on the wielder's current health. While this still falls off at
low HP, ones HP must be very low indeed for this to become a bad weapon.

The numbers could probably use some further tweaking.