Remove some broken local tiles clicking stuff
This has been broken in an unusable way since c3daa31c0c (about two
years) because of the incorrect use of testbits, and even fixing that
these features are still pretty unusable in my testing. They are also now
somewhat redundant with the quiver features. I think something like the
previous behavior of ctrl-click and alt-click could be a good thing to
have, but it needs to be rewritten (perhaps using the actions api) and
restructured to be less redundant. One idea in the long run might be to
have these extra click combos access more quivers, or quivers that are
spell and evocable specific (similar to the launcher quiver).

Given how broken this has been, I'm just removing for now, the main
place to notice a difference will be that ctrl-click on the player will
now not lead to broken/confusing behavior with targeted spells. There's
shift-click to cast/evoke from the new quiver, and also still a
click-to-cast interface in the sidebar for local tiles.