Don't feel safe on hexers trapped in grates (12371, #1651)
If a monster had only non-damaging ranged spells and was trapped in
grates (or water, etc), this code would register that the monster could
not hurt the player. This is narrowly true, but because this function is
used for i_feel_safe, the monster would not count as a threat for the
player under various broader  conditions, notably, autotravel.

The specific buggy case that has come up in several reports is draining
eyes, which don't damage the player but do of course drain mp; in other
circumstances they register as dangerous by having a foe path to the
player. In 0.25, `explore_auto_rest` became default true, and so if the
player autotraveled into los of a trapped draining eye, they would enter
a perpetual rest + drain cycle because i_feel_safe returned true,
running up the clock until they interrupted it. This is even worse for
players who have a force_more on the drain message (e.g. ones using
HDA's rc as in #1651) because it is hard to interrupt the loop. It's
also compounded by the lack of hunger: in 12371 the player got out of
the autoexplore loop when they became hungry, which can't happen any
more.

The change is to construe "hurt" for this function's purposes more
broadly, to include any dangerous ranged spells regardless of whether
they are attacks; it's used only for autoexplore and as a minor check on
tension.  It's possible this will stop autoexplore incorrectly in some
cases I haven't thought of, but it can be fine-tuned from here.

Resolves #1651

(cherry picked from commit 0dbc5c1afff67718724f1e3d2fa08ccf45669cf6)