Some changes to the new Swamp end vault
For nzn_swamp_witches_coven, add deep elf mages to the "wizard" set and
reduce the weighting for spriggan druids, giving their previously double
weight to the mages and to the wizards that previously had half weight.
Druids have a somewhat harder time in this vault since there are so few
trees, so it's best to not place so many of them.

For the inner areas that randomly get shallow water, land, or monsters,
increase the amount of land by 50% and decrease the amount of water.
This vault previously placed an awful lot of water without very much
land. Some tests with the new weight shows that there is still mostly
water with relatively few tiles safe from trees, but it's not going to
be quite so hard to find an area to fight inside the vault. Otherwise
the incentive is to clear nearly all of the vault from outside.

Also, make the inner pentagram walls use special dark stone wall tiles
that more strongly suggest a demonic theme, using magenta for these
walls in console. This is purely a cosmetic change with no affect on
gameplay. Finally, make the central granite statue always use a
demonic-themed tile.