Some minivault adjustments
For nzn_sisyphean_effort and nzn_prometheus_bound, have the placed
corpse on a delayed decay instead of never decaying. For the Prometheus
vault, place a Book of Fire instead of a Book of Flames, since the
former is a more appropriate reward in Shoals and just as thematically
appropriate.

For nzn_life_drinkers, simply clean up some SUBST. This vault could
place as an overflow altar with low weight for D:8-9. In Vaults, ideally
it would be tagged as a vaults room so that it could place successfully
more often. It works fine as a standalone altar vault altar and can
successfully place in Vaults per wizmode tests, so I can save both
potential changes for some time before the 0.26 release.