Make Irradiate check AC and bump damage
Not checking AC was an oversight in the initial implementation that
slowly transmuted into a feature. Well, let's transmute it back.

I've made the damage a bit higher to compensate. Per a suggestion from
hellmonk, I've also reduced the number of dice that are rolled, which
reduces the predictability of the spell a bit.