Add a bit more control to fire_order
This is a bit of a filler commit; I think fire_order needs to be
restructured in a more general way, but for now at least provide some
coarse control over the order of action types when cycling the quiver.
This commit also lets players remove an entire type from the cycle order
(which has been particularly requested for abilities).

Because game start autoquivering uses `cycle()`, this commit also
affects that, so for the people who find it objectionable that Berserk
autoquivers on a Be start, this gives them a way to prevent that.

Eventually I think this would be *much* better as an in-game interface
along the lines of spellbook hide/show.