Unify the two cursors on local tiles X somewhat.
Clicking on a place on the map will move the map cursor to there.
Clicking on the map cursor will travel to there. It would be nice to
unify them entirely, but the problem is that X automatically centers the
map around the map cursor, so moving the mouse will shake around the
view erratically.