Retheme /disintegrate -> /mindburst
Disintegration is a fun effect (everyone loves things blowing up),
but it's always been a weird exception in certain ways. MR (now Will)
was always the stat that resisted debilitating effects "plus Disintegrate",
and there was never any real flavour justification for that. Why could
you resist Disintegrate but not, say, Energy Bolt? It makes it hard for
players to develop intuitions about what affects what.

So let's retheme. Instead of an abstract Disintegration, Mindburst
specifically launches a psychic attack on the target's brain, attempting
to blow them apart from the inside. It should be a bit more intuitive why
this is affected by Willpower (since it's a mental attack), and it still
preserves all the explosive fun of the original.

For the sake of intuition, the new Mindburst does *not* affect mindless
creatures. This mostly shouldn't matter - there aren't that many mindless
creatures that you wouldn't want to disintegrate anyway. (It might be a bit
rough if you're up against an early centaur skeleton, maybe.)

Great Orb of Eyes' Disintegrate spell has been changed to use Mindburst as
well.