Add detailed static targeters for LOS enchantments
To round out the new quiver and collection of static targeters, this
commit adds detail to the static targeters for: Cause Fear, Discord,
Alistair's Intoxication, and Metabolic Englaciation; also added is a
static targeter for the pseudo-spell Drain Life.

Each now highlights the monsters that will be affected (or at least not
outright immune). Additionally, success chance displays have been added,
so that the chance of affecting each target is available.

In an ideal world these spells would be further re-factored so that the
targeter behavior is unified with the casting behavior, the affect logic
encapsulated in beam or zap data, and the different targeting classes
unified in a single targeter_multiposition_enchantment.

For now, I am content with the UI clarity.