Allow static-targeter preview of Confusing Touch
By enabling these two spell flags the recently implemented static
targeter interface Just Works to allow a preview of Confusing Touch
success chances. The highlighting intentionally highlights only the
player, which might make this behavior a bit hard to discover, but it
feels less misleading than "targeting" all visible monsters. The wording
could be improved, but currently the success chance display does not
allow verb customization.

Because this targeter is a bit of a kludge, I haven't made it a default
force targeter.