Re-refactor targeter_multiposition
Commit 464169c878 fixed some of the monster checking that
targeter_multiposition does, and subsequent commits used
targeter_multiposition as a base class for static targeters for various
LOS-enchantment effects.

The implementation of 0242afd26b had some order-of-execution finesse to
ensure that a virtual method called by the targeter_multiposition
constructor was correctly initialized. In that commit and the comments
in it I remarked there was probably a better way. Indeed, the design
smell of the constructor pointed out the following: multiposition and
multimonster targeting, while appearing similar in the ui, are
semantically different. This commit splits up users of the class:
targeter_multiposition is used when the targeted cells depend
non-locally on the monster occupying the cell, and a new targeter class
targeter_multimonster is introduced for targeting monsters in los on a
monster-by-monster basis.

The result simplifies targeter_multiposition nicely.