Improve tremorstone + summons warning checks (12409)
This is a fairly minimal approach, in contrast to the potential for
refactoring: there's a lot of related code that could be consolidated
here, and it looks to me like shoot_through_monster does a bunch of
checks that doesn't need the full beam. Various other cases have the
same code duplication issue, see e.g. b68aede. However, I don't want to
mess with this stuff in a general way during feature freeze.