Make player drain fully proportional and buff it
I originally made max HP drain split power into two components, one used
for a randomized cost and one for a proportional cost. This was to
balance the effect between low HP characters and high HP characters. Low
HP characters get penalized more an effect based on a fixed value
compared to one that's based on a percentage of max HP. Making it fully
percentage-based hurts high HP characters more, but elliptic pointed out
that this is how other max HP reduction effects like frail work. Based
on play-testing, the current drain effect is a bit weak for both monster
attacks and ability drain costs, and we'd like to buff it. Making the
effect fully proportional will let us do this without harming low HP
characters quite so much. Yes, we actually care about the poor felids
(no we don't).

This commit makes the effect fully based on the player's current max HP
and increases its strength by 33%. Now the final HP drained is simply:

1 + div_rand_round(power * max_hp, 750)

with the division being randomly rounded in the usual way. Hence we
always drain 1 MHP if our power was non-zero. Drain power is still 1 +
damage for monster attacks or a fixed cost value for abilities, as
before,

Some quick play testing has shown early wight packs to be a bit more
effective at getting the player into lightred drain and e.g. Ru
Apocalypse to be a bit nastier. If this proves to be not enough, it's
easy to adjust that 750 divisor.