Clean up + simplify "Powers" tab Invocations info
The original form of this message ("piety instead of invoc") was already
kind of confusing, because most invoc gods are affected by *both* piety
and invoc (and this is described for specific powers on the Overview tab
too). In addition, it just left out gods where the "piety" part wasn't
appropriate or was too involved. This commit simplifies the skill
message to only describe the non-impact of Invocations, and shows this
message for all gods that aren't affected by invoc. Contra the
now-removed comment, there would be space to describe any additional
effects that this doesn't cover, but it's not actually that obvious to
me what this comment was talking about that is missing. (For Gozag, the
effect of gold is well-described, and the comment about WJC seems
outdated?)

Resolves #1662