Restore switch-on-autopickup behavior from pre-0.26
When you ran out of one type of ammo, autoswitched to another, and then
picked up the first type, oldquivers would switch you back. It got
removed in the quiver rewrite because the code was kind of messy. This
commit restores that behavior, by popular demand. I'm not super happy
with the implementation (which unfortunately relies on a bunch of weird
legacy code) but after thinking about this for a while, I'm not sure
there's a clean way to implement this behavior.

(cherry picked from commit bd55b0887b59281757b96ef1c3a321a374cbf3d6)