Fix quiver redraw issues on delay interrupt
Depending on the exact timing, a stat redraw sometimes wouldn't happen
immediately after a delay interrupt, leading to incorrectly showing a
quiver as disabled. This commit adds some delay-related checkpoints to
ensure that the quiver info is up-to-date, and forces a stat redraw on
runrest delay end. (This could come up for other interrupted delays?)

This is somewhat delicate to debug/replicate because messaging can force
a stat redraw itself.