Fix a glyph size issue (powerbf)
I *think* what was happening here is an off-by-one error; if the glyph
size is exactly the max advance the stored version was failing to be
doubled, and was short one column; this column was getting overwritten
by an adjacent glyph. Huge thanks to powerbf in #1482 for doing the
bisection + debugging to figure out the source of the problem. I wont'
auto-close that PR as I'd like to get this tested on other OSs.