Don't process click_to_travel minimap clicks in map mode
If you left-clicked the minimap while in map (`X`) mode, this would
start autotravel, and then "continue" it when exiting map mode in any
way including escape. In #1508 it was also reported that this sequence
of actions caused a crash on windows; it isn't surprising to me that
activating click_to_travel without leaving map mode would crash, though
I didn't get it to crash on mac. In any case, this behavior doesn't make
sense (escape should abort), so remove it, and hopefully fix the crash
on windows as well.