Remove some console blinkiness in map view
As far as I can tell, this update_screen call is guaranteed to force an
update on a blank screen, and since it is triggered on every refresh in
map view, it is making map view extra blinky (especially over ssh). I
can't find any negative consequences of just removing it.