Fix webtiles rendering of ctrl keys
I honestly have no idea how these were displaying in console, but they
were, so I assumed they would work in webtiles as well. But something
low-level that I haven't found must have been doing the actual rendering
in a way that didn't convert them to readable strings in webtiles. This
is a quick patch to render them correctly in the macro menus.