Simplify Tornado interactions with distance movement
Prior to this commit, if the player moved by translocating (or Ba hop)
multiple tiles while under the effects of tornado, the tornado was
cancelled, without requiring cooldown, UNLESS the translocation landed
the player in a tornado cloud, in which case the tornado duration was
decremented by a translocational penalty.

No change to duration was incurred on taking a passage of golubria or
transporter.

On taking stairs the tornado was cancelled with no cooldown.

This commit changes the behavior to: blinknado is ok, but all other long
distance changes end tornado and give the cooldown.

Closes #1611