Adjust the charging of Majin-Bo's hp cost
Don't send it through ouch, it should be treated like an ability hp
cost, not damage (so no harm scaling!). Also, dec the hp at the same
place as the mp cost of the spell, for appropriate sequencing of the
effect and to give the player a preview of the cost while targeting.

Closes #1578

(cherry picked from commit 0fe236cf81fa067bbd5ccc5c028db979305c62a7)