Don't update constriction when testing door seals
The vault warden AI check to see if it should seal a door
was using move_to_pos to evaluate the tension of actor moves, but
move_to_pos handles updating constriction information of the
constrictor. The result was that if the check needed to test if pushing
something through a door in a way that would break constriction was a
good idea, the constriction would actually break, even if the monster
ultimately decided to not close the door!

Closes #1566

(cherry picked from commit 897d3da516b74d2ab7e49b61dc4f072745eeb5fa)