Use erase() for some ncurses screen clears
This is an attempt to mitigate visual flashing seen with some ncurses
versions; `clear` calls `clearok` which is more likely to actually show
a blank screen for a short amount of time to the user. Nonetheless, I'm
not sure why this only happens for some versions or if there's something
more going on, so this is a bit experimental and may need more tuning.