New spell: Animate Armour
L4 Earth/Summoning. Call forth the spirit of your armour, bringing forth
a slow-moving ally that hits harder the heavier your armour is. Casting
this in robes creates a 40 HP, 4 AC ally that hits twice for 4 damage.
In plate armour, you get a 40 HP, 20 AC ally that hits twice for 60 damage.
Wham, pow! (Summon limit is 1, as with CCF.)

The design goal here is to help provide some incentives for heavier armour
characters to delve into spellcasting. Historically, spells like Haste and
Phase Shift tempted players in this direction. Those spells had their own
problems, but the 'hybrid' skilling they encouraged was fun. This spell
will still be inappropriate for many characters, but if it brings a few
'melee' characters to dabble in spellcasting or 'caster' characters to
put on heavier armour, I think that'll be great.

Numbers are very much a first draft. I can easily imagine that the damage
should scale differently or that HP should vary by armour type, for example.

Many thanks to everyone who helped with this. Special thanks to hellmonk
for balance suggestions, ebering for code and balance suggestions, mesh and
rojja for German advice, and canine for the final name.