Reduce net trap trigger chance
Since making net traps not run out of ammo, they've become more vicious
than was intended. To mitigate this, give them a 2/3 chance of *not*
triggering when a monster enters them.

It'd be interesting to make them require LoF to hit the player, but
that's an experiment for another day.