New spell: Manifold Assault (L5 Translocations)
Manifold Assault launches melee attacks with against 2-4 random
enemies in LOS, depending on spellpower. It's themed as warping space
so that one attack can, paradoxically, hit many enemies.

This spell is intended to give melee-focused characters more reason to
consider investing in spellcasting, It's fun to try to skill toward a
'hybrid' build, and a spell that mostly scales off your melee attack,
rather than your spellpower, helps encourage that. This being a L5
spell helps avoid it being a truly free 'gimme', and allows it to hit
many enemies at once.

The lack of targeter should hopefully avoid it being annoying to use,
by making it just a single press to cast. It also gives the player
more reason to keep normally hitting things in melee sometimes,
when it's important to kill a specific target.

Cleaving doesn't work with this spell since it seems overly confusing,
and since reach and riposte don't work with it either. (Thus, it's
'fair'.) All stabs are also currently banned, though I do wonder if
we might want to allow non-sleep stabs.