First pass conversion of webtiles to asyncio
This implements the key domino pieces of trying to make the core
webtiles server use coroutines, rather than old-style Tornado callbacks.
The old implementation is getting increasingly painful to deal with and
maintain, relies on deprecated/removed Tornado API pieces, and blocks in
a lot of cases that it shouldn't. This commit deals with the main parts
of webtilies server by converting everything necessary to coroutines.
There are still parts of the old implementation present that are not
ideal (e.g. connection.py still has blocking socket code that does even
cause issues in live servers).

I don't currently know how exactly this should be integrated into the
repository; needless to say that it breaks python 2 compatibility fully,
and probably even python 3.4 compatibility. So as it stands this commit
is not ready for merging, and maybe the old webtiles code needs to be
preserved in the repo somehow in a way that allows servers that cannot
run py3 (e.g. cao, as it stands) to still use it.