Change Cerbov's summon
Summon Greater Demon is of course something a lord of Pandemonium should
be able to do, but it seems a bit silly flavour-wise for the firey lord
of Pandemonium to summon a blizzard demon. This replaces Cerebov's SGD
with the buffed Fire Summon spell of Asmodeus, so he now only calls
Balrugs, Hellions, and Brimstone Fiends.

It's not completely clear that Cerebov even needs summons, since he's
the Fire Storm monster, but summons do prevent a player from isolating
the boss monster completely and Cerebov seems to work well with them, so
let's try it.