Adjust and derandomize shield delay penalty
The old formula was approximately an extra min(1d adjusted shield penalty, 1d adjusted shield penalty) delay. However, this was calculated at a scale of 20 to
allow die with penalty less than 1, since for a human-sized player the
adjusted shield penalty of a buckler at 0 skill is 1. With scaling, the
formula was:

min ( 1 d (20 * penalty), 1 d (20 * penalty) ) / 20

randomly rounded.

As a result, at 0 skill a buckler did not have a delay penalty of 0.1,
but in practice was something that was 0 with a pretty high probability.
To get a very good chance of a delay penalty of 0, 7 skill was
sufficient for a kite shield, and 12 for a tower shield.

7 skill costs a human 1,400 skill points; 12 skill costs a human 4,200
skill points.

2 hand maces (excluding dire flails), pole arms, and
swords, all require 20 skill (or more) for min-delay; at the probability
of penalty achieved for the above shield skill values, this is a better
comparison point than the skill required for min+0.1 delay. 20 weapon
skill costs 13,950 skill points, compared to 9,600 for 16 skill and the
7 shield skill; the difference in base damage is not
sufficient compensation for enough skill points to raise a 0-apt skill to 10!

This commit changes the formula to something less random and stiffer,
namely:

random_round(penalty)

At 0 skill, a buckler has a flat constant delay penalty of 0.1,
eliminated at 4 skill. At 10 skill a kite shield gives a delay penalty
of exactly 0.1, and the old probability is achieved at a skill level
between 14 and 15. To get a delay penalty of 0.1 with a tower shield,
20 skill is required.

With these skill training targets, achieving a delay of min+0.1 for a 16
skill 1-hander while using a kite shield costs 11,000 skill points,
compared to the 10,800 skill points for 18 skill to get the same swing
speed with a 20 skill 2-hander. For min-delay, the 16 skill 1-hander +
kite costs 15,250, compared to 13,950 for a 20 skill 2-hander.

This is a better alignment of strategic costs, which will hopefully
make the "1h vs 2h" a more interesting strategic decision based on game
context.

The increase to broad axe delay in 81720017eb when combined with this
change might make broad axes too expensive; that is easily fixed if it
turns out to be the case.