Don't do zone filling in the abyss
Unlike normal builder scenarios, connectivity checks are only rarely
triggered in the abyss, by the placement of a minivault it seems. (I'm
surprised they happen at all, and wonder what they accomplish.) In this
case, it seems wrong to do any fills, as the abyss is supposed to allow
temporarily disconnected zones.

In mantis 12442, 12416, as well as #1680, it was reported that this
scenario could lead to errors, because the fill code assumed that no
items had been placed yet. As far as I know this is mostly true, just
not for the abyss, which does most of its building before minivaults
have the chance to place. I've also added a guard against filling over
items, even though the main fix in this commit prevents the specific
case from coming up. Interfering items just get nuked, to prevent vetos.

Resolves #1680.