Don't leak stash info with Gozag gold detection (12413)
This makes gold detection work more like item detection, in that the
detected gold does not appear in stash search as it did previously. This
is a bit unfortunate, but this is the lone edge-case remaining where a
stash pile is revealed partially. It shows that the detected thing is
gold versus an item for a gnoll on the map, which suffices.

(cherry picked from commit f9d094ba9912b00531782455c027dcf27446a2e6)