Use a name for ' ' when a name is called for
This is a printable char, but isn't apparent in most circumstances when
it would be printed directly, so ' ' is not useful as the character
description. Therefore, use "Space". Impacts only menu hotkeys, and
probably only in the macro menu.

(cherry picked from commit dd83002478a0ce41aafb116edbfb50769263d21c)