Add a better UI range check for minor destruction (12451)
Minor destruction did not have a range check except on activation,
leading to bad behavior with autofight_throw=true when quivered.  This
commit follows the same strategy in 0579ed41f9 for rolling charge. The
background problem is that the ability API is still somewhat
impoverished, and we don't want to fully disable the ability with no
enemy in range (because there is no way to force-activate it equivalent
to casting `Z`). However, for quiver and autofight purposes the quivered
ability needs to be marked as disabled in this scenario. Eventually it
would be nice if spells and abilities had a more uniform interface for
these issues.