Adjust some empty quiver cycle messaging
This gets rid of the vestiges of oldquiver error messages, and rewrites
things to use the newquiver code in a more straightforward way. This
removes some of the hints that the old errors gave about inscriptions
etc, but I think the simplification her targets everything that will be
likely to apply to newer players who need the hints in the first place.